requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
ecdsa==0.18.0
frozendict==2.4.0
idna==3.6
lxml==5.1.0
pyasn1==0.5.1
PyLD==2.0.3
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
urllib3==2.1.0
